% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledresampling-systematic-given.R
\name{CR_systematic_given}
\alias{CR_systematic_given}
\title{Coupled Resampling: systematic, given vector of ancestor}
\usage{
CR_systematic_given(xparticles1, xparticles2, normweights1, normweights2,
  parameters = list(), ancestors_ref, uniforms)
}
\value{
A vector of ancestors
}
\description{
This function performs systematic resampling on the second system
}

