% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledresampling-transport.R
\name{CR_transport}
\alias{CR_transport}
\title{Coupled Resampling: transport}
\usage{
CR_transport(xparticles1, xparticles2, normweights1, normweights2,
  parameters = list(epsilon = 0.1, desired_alpha = 0.9))
}
\value{
Two vectors of ancestors, column-binded in a matrix.
}
\description{
This function performs coupled resampling based on transport
The parameters are given as a list with keys
* epsilon, which determines the approximation in the solution of the transport problem
* and desired_alpha, which determines implicitly the number of iterations performed by the iterative
algorithm.
}

