% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupledresampling-transport-given.R
\name{CR_transport_given}
\alias{CR_transport_given}
\title{Coupled Resampling: transport, given vector of ancestor}
\usage{
CR_transport_given(xparticles1, xparticles2, normweights1, normweights2,
  parameters = list(epsilon = 0.1, desired_alpha = 0.9), ancestors_ref,
  uniforms)
}
\value{
A vector of ancestors
}
\description{
This function performs tranasprt resampling on the second system
}

