% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_storeall.R
\name{particle_filter_storeall}
\alias{particle_filter_storeall}
\title{Particle Filter}
\usage{
particle_filter_storeall(nparticles, model, theta, observations, randomness)
}
\arguments{
\item{nparticles}{number of particles}

\item{model}{a list representing a model, for instance as given by \code{\link{get_ar}}.}

\item{theta}{a parameter to give to the model functions}

\item{observations}{a matrix of observations of size datalength x dimension(observation)}

\item{randomness}{an object storing all the variables useful to initialize and propagate the particles,
as generated by a model$generate_randomness function}
}
\value{
A list with entries ll (loglikelihood estimator), xhistory, ahistory, whistory
(lists containing all the generated x-locations, ancestors and weights)
}
\description{
Runs a particle filter, storing all the generated variables
}

